backend/requirements.txt
fastapi
uvicorn
python-barcode
Pillow
python-multipart 